final price long

+++ app/src/main/java/com/partsystem/partvisitapp/feature/report_factor/offline/model/FactorHeaderDbModel.kt
@@ -8,7 +8,7 @@ data class FactorHeaderDbModel(
     val patternName: String?,
     val persianDate: String,
     val createTime: String,
-    val finalPrice: Long,
+    val finalPrice: Double,
     val hasDetail: Boolean,
     val sabt: Int
 )
